# Conflicts:
#	src/main/java/com/example/PizzUMBurgUM/entidades/Cliente.java
#	target/classes/com/example/PizzUMBurgUM/entidades/Creacion$CreacionBuilder.class
#	target/classes/com/example/PizzUMBurgUM/entidades/Creacion.class

+++ src/main/java/com/example/PizzUMBurgUM/controladores/ControladorTicket.java
@@ -14,7 +14,7 @@ public class ControladorTicket {
     public ControladorTicket(CreacionServicio creacionServicio) {
         this.creacionServicio = creacionServicio;
     }
-    public TicketDto obtenerTicket(@PathVariable Integer idCreacion) {
+    public TicketDto obtenerTicket(@PathVariable Long idCreacion) {
         return creacionServicio.generarTicket(idCreacion);
     }
 }
