+++ src/main/java/com/example/PizzUMBurgUM/dto/TicketDto.java
@@ -2,7 +2,7 @@ package com.example.PizzUMBurgUM.dto;
 import java.util.List;
 
 public record TicketDto(
-        Integer idCreacion,
+        Long idCreacion,
         String clienteEmail,
         List<DetalleDto> detalle,
         int total
